+++ src/scipp/core/__init__.py
@@ -32,11 +32,14 @@ DTypeError.__doc__ = 'Inappropriate dtype.'
 
 from .._scipp.core import get_slice_params
 
-from .sizes import _make_sizes
+from .dimensions import _make_sizes, _rename_variable, _rename_data_array, _rename_dataset
 
 setattr(Variable, 'sizes', property(_make_sizes))
 setattr(DataArray, 'sizes', property(_make_sizes))
 setattr(Dataset, 'sizes', property(_make_sizes))
+setattr(Variable, 'rename', _rename_variable)
+setattr(DataArray, 'rename', _rename_data_array)
+setattr(Dataset, 'rename', _rename_dataset)
 
 from .bins import _bins, _set_bins